+++ spec/examples/expected_fail_in_must_when_type_is_last_advanced.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class ExpectedFailInMustWhenTypeIsLastAdvanced < Actor
+  input :per_page,
+        must: {
+          be_in_range: {
+            is: -> per_page { per_page >= 3 && per_page <= 9 },
+            message: -> value:, ** { "Wrong range (3-9): #{value}" }
+          }
+        },
+        type: Integer
+
+  def call; end
+end